Login Improved + Updated SQL Script
I have done what Bohus said, authentication method now return the ID of the person not the password. So now the credentials are being compared on the client side, not in the app. Also i reduces the risk where 'TEST' and 'test' would let you login, so now it is sensitive capitals. Also, you need to run the SQL Script in order this to work. I put there a script to create the procedure and a script to ALTER the procedure, for those who have it already.